+++ srcs/read_input/edit_input.c
@@ -1,87 +1,7 @@
 # include <libft.h>
-# include "utils.h"
-# include "abstract_list.h"
 # include "read_input/editor/editor.h"
 # include "read_input/event_callbacks/event_callback_def.h"
 # include "read_input/termcap/init_deinit.h"
-
-void move_start(t_editor *ed)
-{
-	t_vec2i	vec;
-
-	vec = get_cursor_vector(ed);
-	while (vec.y > 0)
-	{
-		ft_putstr(ed->term->move_up);
-		vec.y--;
-	}
-	ft_putstr(ed->term->move_cursor_begining);
-}
-
-void restore_old_cursor_position(t_vec2i old_pos, t_vec2i new_pos, t_term *term)
-{
-	move_cursor(vec2i_sub(old_pos, new_pos), term);
-}
-
-void print_command_string(char const *prompt, t_string *cmd_str
-	, size_t term_width)
-{
-	char	**lines;
-	char	*str;
-	char	*tmp;
-	size_t	u;
-
-	tmp = get_string_from_list(cmd_str);
-	str = ft_strjoin(prompt, tmp);
-	free(tmp);
-	lines = ft_strsplit(str, '\n');
-	u = 0;
-	while (lines[u] != NULL)
-	{
-		ft_putstr(lines[u]);
-		if (ft_strlen(lines[u]) % term_width == 0)
-			ft_putchar('\n');
-		if (lines[u + 1] != NULL)
-			ft_putchar('\n');
-		u++;
-	}
-}
-
-void clear_line(EV_CB_ARGS)
-{
-	move_start(ed);
-	ft_putstr(ed->term->clear_line);
-	ed->tmp = get_cursor_vector(ed);
-}
-
-void put_line(EV_CB_ARGS)
-{
-	char		*line;
-
-	print_command_string(ed->prompt, ed->string, ed->term->width);
-	line = get_string_from_list(ed->string);
-	ed->old_position = ed->cursor_position;
-	ed->cursor_position = ft_strlen(line);
-	free(line);
-}
-
-void refresh_line(t_editor *ed)
-{
-	if (ed->need_refresh == true)
-	{
-		ed->need_refresh = false;
-		clear_line(ed);
-		put_line(ed);
-		restore_old_cursor_position(get_cursor_vector(ed), ed->tmp, ed->term);
-		ed->cursor_position = ed->old_position;
-	}
-}
-
-void free_editor(t_editor *ed)
-{
-	free_string(ed->string);
-	free(ed->prompt);
-}
 
 char *edit_input()
 {
