+++ srcs/read_input/editor/editor.c
@@ -34,7 +34,7 @@ t_editor	*init_editor()
 	return (new);
 }
 
-void	add_to_string(t_editor *ed, char c)
+void	add_to_string(EV_CB_ARGS, char c)
 {
 	t_string *new;
